refactor: split the plan editor's drafts and containment into their own modules
